+++ src/App.scss
@@ -3,4 +3,5 @@
   display: flex;
   flex-direction: column;
   justify-content: center;
+  align-items: center;
 }
