+++ src/App.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import './App.scss';
 
-import { PostInfo } from './PostInfo/PostInfo';
+import { PostList } from './components/PostList/PostList';
 
 import postsFromServer from './api/posts';
 import commentsFromServer from './api/comments';
@@ -9,14 +9,14 @@ import usersFromServer from './api/users';
 
 const preparedPosts: Post[] = postsFromServer.map(post => ({
   ...post,
-  comments: commentsFromServer.filter(c => c.postId === post.id) || null,
+  comments: commentsFromServer.filter(c => c.postId === post.id) as Comment[],
   user: usersFromServer.find(u => u.id === post.userId) || null,
 }));
 
 const App: React.FC = () => (
   <div className="App">
     <h1>Static list of posts</h1>
-    <PostInfo preparedPosts={preparedPosts} />
+    <PostList preparedPosts={preparedPosts} />
   </div>
 );
