+++ src/components/PostList/PostList.scss
@@ -1,6 +1,7 @@
 .post-list {
   padding: 0;
   list-style: none;
+  max-width: 1024px;
 
   &__item {
     margin-bottom: 20px;
